refactor: move slots logic to own file
